backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg[binary]==3.2.12
pydantic-settings==2.3.4
redis==5.0.7
fakeredis==2.24.1
python-jose==3.3.0
aio-pika==9.4.1
websockets==12.0
bcrypt==4.1.2


